GRAL-2614 refactor tests
